feat: Allow users to set Pod DNS config
This change is triggered by https://github.com/external-secrets/external-secrets/issues/4824

Added `podDnsPolicy` and `podDnsConfig` to the Helm chart to allow
customization of DNS policy and configuration for pods. This enables
users to specify DNS settings such as nameservers, search domains,
and options.

Since `bitwarden-sdk-server` only really has one purpose, for users that use the Bitwarden server,
it might be simpler to set the DNS config one time and not need to set `apiURL` + `identityURL`.
Relying on defaults

Signed-off-by: Yasser Saleemi <yassersaleemi@gmail.com>